requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Async Processing
celery==5.3.4
redis==5.0.1

# AI Integration
openai==1.3.5
anthropic==0.7.10

# Web Scraping & Analysis
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Email & Communication
sendgrid==6.10.0
mailgun==0.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.39.1

# API Clients
googlemaps==4.10.0
linkedin-api==2.0.0